fuel stored as decimal on back end

+++ Data/DigState.cs
@@ -13,7 +13,7 @@ namespace digsite.Data
         public int DigStateId { get; set; }
         public int PlayerId { get; set; }
         public int Depth { get; set; }
-        public int Fuel { get; set; }
+        public decimal Fuel { get; set; }
         public byte IsPaused { get; set; }
 
         public Player Player { get; set; }
